requirements.txt
openai
python-dotenv
opencv-python
mediapipe
fastapi
streamlit
pillow